requirements.txt
# RAG Core (FIXED)
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.10
langchain-openai==0.1.17

# Pytorch
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# Embeddings / VectorStore
sentence-transformers==2.7.0
faiss-cpu==1.8.0
huggingface-hub>=0.23.0

# Graph-RAG
neo4j==5.24.0
scikit-learn==1.5.0
numpy>=1.26

# PDF Processing
pymupdf==1.24.7
pypdf==5.1.0

# ChatGPT client & environment
openai>=1.33.0
python-dotenv==1.0.1
requests==2.29.0

# Streamlit UI (RAG-Naive)
streamlit==1.35.0
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
base64io>=1.0.3

# Utils
tqdm

# Server
fastapi==0.104.1
uvicorn==0.24.0